Move some timeline code
